+++ client/src/main/java/com/fastcomments/model/PickFCommentPublicCommentFieldsKeys.java
@@ -182,7 +182,7 @@ public class PickFCommentPublicCommentFieldsKeys {
   public static final String SERIALIZED_NAME_VIEW_COUNT = "viewCount";
   @SerializedName(SERIALIZED_NAME_VIEW_COUNT)
   @javax.annotation.Nullable
-  private Double viewCount;
+  private Long viewCount;
 
   public static final String SERIALIZED_NAME_REQUIRES_VERIFICATION = "requiresVerification";
   @SerializedName(SERIALIZED_NAME_REQUIRES_VERIFICATION)
@@ -688,7 +688,7 @@ public class PickFCommentPublicCommentFieldsKeys {
   }
 
 
-  public PickFCommentPublicCommentFieldsKeys viewCount(@javax.annotation.Nullable Double viewCount) {
+  public PickFCommentPublicCommentFieldsKeys viewCount(@javax.annotation.Nullable Long viewCount) {
     this.viewCount = viewCount;
     return this;
   }
@@ -698,11 +698,11 @@ public class PickFCommentPublicCommentFieldsKeys {
    * @return viewCount
    */
   @javax.annotation.Nullable
-  public Double getViewCount() {
+  public Long getViewCount() {
     return viewCount;
   }
 
-  public void setViewCount(@javax.annotation.Nullable Double viewCount) {
+  public void setViewCount(@javax.annotation.Nullable Long viewCount) {
     this.viewCount = viewCount;
   }
